output: Split wayland::output
