+++ src/desktop/window.rs
@@ -1,10 +1,10 @@
 use crate::{
     backend::renderer::{utils::draw_surface_tree, ImportAll, Renderer},
     desktop::{utils::*, PopupManager, Space},
+    output::Output,
     utils::{user_data::UserDataMap, IsAlive, Logical, Physical, Point, Rectangle, Scale},
     wayland::{
         compositor::with_states,
-        output::Output,
         shell::xdg::{SurfaceCachedState, ToplevelSurface},
     },
 };
