+++ src/lib.rs
@@ -56,6 +56,7 @@ pub mod backend;
 #[cfg(feature = "desktop")]
 pub mod desktop;
 pub mod input;
+pub mod output;
 pub mod utils;
 #[cfg(feature = "wayland_frontend")]
 pub mod wayland;
